requirements.txt
transformers
datasets
mlflow
scikit-learn
pandas
pyyaml
gradio
prometheus-client
dvc
